Intro - Generate project and make a change
In terminal, run:
  swift package generate-xcodeproj
then open the auto-generated project file, and navigate to Sources/server/main.swift and make a change to the print statement.

To run, choose scheme with the black icon (not the yellow briefcase) and device ‘My Mac’, and press Cmd+R to see the hello message shown in the Xcode console.

Note: NEVER make changes to your XCode project, always let SPM do it for you.